refactor(api): [#143] extract mods for API testing
Code for API testing have been reorganized.
