+++ src/apis/mod.rs
@@ -1,2 +1,2 @@
-pub mod server;
 pub mod routes;
+pub mod server;
