+++ src/jobs/mod.rs
@@ -1,5 +1,5 @@
 pub mod http_tracker;
 pub mod torrent_cleanup;
 pub mod tracker_api;
-pub mod udp_tracker;
 pub mod tracker_apis;
+pub mod udp_tracker;
